# cmd/api/openapi.yaml
openapi: 3.0.3
info:
  title: Job Scheduler Platform API
  version: 1.0.0
  description: |
    Durable background-job platform: submit jobs over HTTP, run them now, later, or on cron schedules.
    Webhook delivery with retries, exponential backoff, a dead-letter queue, per-client rate limits,
    and Prometheus observability.
  contact:
    name: Job Scheduler Platform
    url: https://github.com/BurhaanAshraf/job-scheduler-platform
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:4000
    description: Local development server
  - url: https://api.example.com
    description: Production server (replace with actual domain)

security:
  - BearerAuth: []

paths:
  /healthz:
    get:
      summary: Health check
      description: |
        Returns 200 if both PostgreSQL and Redis are reachable, 503 otherwise.
        This endpoint is public (no authentication required).
      operationId: healthz
      tags:
        - System
      responses:
        '200':
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok]
                    example: ok
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /metrics:
    get:
      summary: Prometheus metrics
      description: |
        Exposes Prometheus metrics in text format. Public endpoint.
      operationId: metrics
      tags:
        - System
      responses:
        '200':
          description: Metrics in Prometheus exposition format
          content:
            text/plain:
              schema:
                type: string
                example: |
                  jobs_submitted_total 42
                  jobs_completed_total 40
                  jobs_failed_total 2
                  queue_depth 5

  /v1/jobs:
    post:
      summary: Submit a new job
      description: |
        Creates a job and enqueues it for execution. If `run_at` is in the past, the job is
        enqueued immediately; otherwise it's scheduled for the future.
      operationId: createJob
      tags:
        - Jobs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreatedResponse'
        '200':
          description: Idempotent replay - existing job returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobCreatedResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency key conflict (same key, different payload)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limited
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until next request allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List jobs
      description: |
        Returns a paginated list of jobs. Supports filtering by status.
      operationId: listJobs
      tags:
        - Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, scheduled, running, done, dead, cancelled]
          description: Filter by job status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum number of jobs to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of jobs to skip
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/jobs/{id}:
    get:
      summary: Get a job by ID
      operationId: getJob
      tags:
        - Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID
      responses:
        '200':
          description: Job found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Cancel a job
      description: |
        Cancels a job if it's still pending or scheduled. Returns 409 if the job
        is already running, done, or dead.
      operationId: cancelJob
      tags:
        - Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID
      responses:
        '204':
          description: Job cancelled
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job cannot be cancelled in current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/jobs/{id}/retry:
    post:
      summary: Retry a dead job
      description: |
        Re-queues a job that has exhausted all attempts (status=dead).
        Resets attempt count to 0 and sets status to scheduled with run_at=now.
      operationId: retryJob
      tags:
        - Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Job ID (must be in dead status)
      responses:
        '200':
          description: Job re-queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Only dead jobs can be retried
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/dead-letters:
    get:
      summary: List dead-lettered jobs
      description: |
        Returns a paginated list of jobs that have exhausted all retry attempts.
      operationId: listDeadLetters
      tags:
        - Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of dead-lettered jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/cron-jobs:
    post:
      summary: Create a cron job
      description: |
        Creates a recurring job definition. The scheduler will spawn job instances
        according to the cron expression.
      operationId: createCronJob
      tags:
        - Cron Jobs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCronJobRequest'
      responses:
        '201':
          description: Cron job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CronJobCreatedResponse'
        '400':
          description: Invalid request (bad cron expression or job template)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/cron-jobs/{id}:
    patch:
      summary: Enable or disable a cron job
      description: |
        Disabling stops future instances from being spawned without deleting history.
      operationId: updateCronJob
      tags:
        - Cron Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          description: Cron job ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCronJobRequest'
      responses:
        '204':
          description: Cron job updated
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cron job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        API key authentication. Generate a key via the CLI tool, then include it as:
        `Authorization: Bearer <raw-key>`

  schemas:
    CreateJobRequest:
      type: object
      required:
        - type
        - payload
        - run_at
        - max_attempts
        - idempotency_key
        - callback_url
      properties:
        type:
          type: string
          maxLength: 128
          description: Job type identifier
          example: "send-email"
        payload:
          type: object
          description: JSON payload to send to callback_url
          example:
            to: "user@example.com"
            subject: "Hello"
        run_at:
          type: string
          format: date-time
          description: When to execute the job (RFC3339). Past = immediate.
          example: "2026-10-04T00:00:00Z"
        max_attempts:
          type: integer
          minimum: 1
          maximum: 100
          description: Maximum delivery attempts
          example: 3
        idempotency_key:
          type: string
          maxLength: 128
          description: Unique key for idempotent submission
          example: "order-12345-confirmation"
        callback_url:
          type: string
          format: uri
          description: |
            HTTP(S) URL to POST the payload to. Must be a public URL (no localhost,
            private IPs, or cloud metadata endpoints).
          example: "https://api.example.com/webhook"

    JobCreatedResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Job ID
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"

    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        type:
          type: string
          example: "send-email"
        payload:
          type: object
          example:
            to: "user@example.com"
            subject: "Hello"
        status:
          type: string
          enum: [pending, scheduled, running, done, dead, cancelled]
          example: "done"
        run_at:
          type: string
          format: date-time
          example: "2026-10-04T00:00:00Z"
        attempts:
          type: integer
          minimum: 0
          example: 1
        max_attempts:
          type: integer
          minimum: 1
          example: 3
        idempotency_key:
          type: string
          example: "order-12345-confirmation"
        callback_url:
          type: string
          format: uri
          example: "https://api.example.com/webhook"
        last_error:
          type: string
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          example: "2026-10-03T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-10-03T12:00:05Z"
        queue_generation:
          type: integer
          minimum: 1
          example: 1

    CreateCronJobRequest:
      type: object
      required:
        - cron_expression
        - job_template
      properties:
        cron_expression:
          type: string
          description: Standard 5-field cron expression (minute hour day month weekday)
          example: "0 2 * * *"
        job_template:
          type: object
          required:
            - type
            - payload
            - max_attempts
            - callback_url
          properties:
            type:
              type: string
              maxLength: 128
              example: "daily-report"
            payload:
              type: object
              example:
                report_date: "{{.RunAt}}"
            max_attempts:
              type: integer
              minimum: 1
              maximum: 100
              example: 3
            callback_url:
              type: string
              format: uri
              example: "https://api.example.com/report-webhook"

    CronJobCreatedResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: Cron job ID
          example: 42

    UpdateCronJobRequest:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: Enable or disable the cron job
          example: false

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "INVALID_REQUEST"
            message:
              type: string
              example: "callback_url must be a valid absolute URL"

  responses:
    Unauthorized:
      description: Unauthorized - missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until next request allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'